requirements.txt
pyserial
tk
Pillow